diagnostics: Reword diagnostic mojom comments
This CL changes the wording around the dependency between
GetAvailableRoutines and the availability of a concrete routine.
The current comment sounds like a necessity that GetAvailable
Routines is necessary to be called before an actual routine is
available, while it is actually a sufficiency.

This CL has a counterpart that applies the same changes to the shadow
of the mojom file in the chromium repo.

BUG=b:245293243
TEST=None

Change-Id: Icc1d5008a034de332c47180bb5792aa0855097d7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3974529
Reviewed-by: Denny Huang <dennyh@google.com>
Reviewed-by: Wei-Luan Wang <weiluanwang@google.com>
Tested-by: Bastian Kersting <bkersting@google.com>
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Commit-Queue: Bastian Kersting <bkersting@google.com>