secanomalyd: memfd execution & baseline UMA metric
Adds the baseline UMA metric for memfd execution, which records the
number of memfd_create system calls that succeed. Also adds the metric
for memfd execution attempt detection. Both metrics are added as enums
to the existing ChromeOS.SecurityAnomaly metric.
To detect the base line condition, a new Parser is added that detects
syscalls in the audit log. The daemon then parses the line to identify
whether the record is a memfd_create syscall that succeeded.
Updates the test data for audit log reader to account for detecting the
syscall type in the audit logs.

BUG=b:255420492
TEST=All memfd_create syscalls are correctly identified. UMA metrics are
emitted to chrome://histograms.
TEST=All secanomalyd tests pass.

Cq-Depend: chromium:3988386
Change-Id: Ibf2ec3eeb2a25d6bc8bb7496c4a772c0c95d2bca
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3976871
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Jeff Xu <jeffxu@google.com>
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Tested-by: Shervin Oloumi <enlightened@chromium.org>