metrics: Fix a rare reporting deadlock when fork
If a process have multiple threads, and one of the threads forks while
the other holds flock to the metrics file, the flock is not going to be
implicitly cleared upon closing the fd in the thread of the original
process. The reason is that fork()'ed process clones the file descriptors
of the parent process, thus adding a holder reference to it. That means
that parent process closing the descriptor only reduces the reference
count, but not closes it. That will keep the flock forever lingering
until the forked process doesn't finish, thus causing a deadlock on
reporting. To fix the problem, the patch adds an explicit unlock on the
problematic code paths (one of the code paths does already have the
unlock), and to make doubly sure, make the metrics file to be opened
with O_CLOEXEC, that will force the fd closure in the forked process
upon exec.

BUG=b:241083140
TEST=There is no reliable repro, so only hope :D

Change-Id: I413ea1ccac61039bcaa3b36fa92c972587e49fc7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3977548
Commit-Queue: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>
Tested-by: Daniil Lunev <dlunev@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@chromium.org>
Reviewed-by: Jong Ahn <jongahn@chromium.org>