shill: Implement stable slot ID for EAP
The underlying PKCS#11 slot ID is not stable, meaning that it is
incorrect to assume that system and user slot always use slot ID
0 and 1 respectively. On specific cases, these values might be
changed. For these cases, shill was not able to correct its slot
ID. Although currently the behavior is fine as NetworkCertMigrator
in Chrome is fixing the mismatches.

In order to support stable slot ID Passpoint, which is not handled
by the cert migrator, this patch implemented the stable slot ID
within shill. The implementation uses underlying shill storage.

Also see: go/cros-stable-cert-id

BUG=b:255445416
TEST=units
TEST=tast run <DUT> network.StableShillEAPSlot;
     (test not yet landed)
TEST=tast run <DUT> network.VPNConnect.wireguard

Change-Id: Ie4161956dd6e70582bac48f1b1fa9b7e5c865199
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4008000
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Michael Ershov <miersh@google.com>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Pavol Marko <pmarko@chromium.org>