shill: vpn: Change GetIPProperties() to return std::unique_ptr
This patch changed the interface of GetIPProperties() to return
std::unique_ptr, and modified the function name to GetIPv4Properties().

This change is to support the IPv6-only VPN case. Shill will be able to
connect to a VPN with only an IPv6 property by changing the IPv4
property field to be optional.

BUG=b:251686186
TEST=unit: FEATURES="test" emerge-$BOARD shill

Change-Id: I6cbc4bd27db1b667663b5eb3332f15bc6e0bb162
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3955028
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Arisa Tsuji <arisai@google.com>
Tested-by: Arisa Tsuji <arisai@google.com>
Reviewed-by: Jie Jiang <jiejiang@chromium.org>