cryptohome: Route all verification through CredentialVerifier instances
This change eliminates the VerifyWithAuthFactorAsync code in the auth
block utility, and instead wraps the call to the FP service in a
CredentialVerifier implementation. So on one hand, this cuts down some
of the AuthBlockUtility usage.

On the other hand, it also ramps it up somewhat, because it moves the
credential verifier factory into said utility. It takes the "auth block
utility knows how to make all the type-specific blocks" and generalizes
that to verifiers, as well.

This also introduces another complexity, specifically due to the
weirdness of legacy fingerprint as a block-less, label-less verifier. To
make this work this change also adds support for having "label-less
verifiers" which can be looked up using the auth factor type instead of
a label. This ends up being a bit gross, but it seemed like a better
decision than making legacy FP a completely special case.

BUG=b:247812443
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ib21d85776c33e53a39a86ba7b61ea3c3947e15a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3994543
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>