vm_tools: cros_im: automate tests as a part of building and packaging
Previously, cros-im GTK3 unit tests must be run manually. This CL
adds these tests to meson.build so that these tests can be run using
the 'meson test' command. More importantly, this allows kokoro to
recognise the tests and run them automatically at the end of every
new package build.

These tests require a wayland server to run, which is not always
available (e.g. on kokoro VMs). To solve this, the test script is
modified to accept an extra argument which will make the script
run the test with xvfb and weston. Note that weston does not work on
armhf so tests have been disabled on armhf for the time being
(b/257867792).

BUG=b:251299561, b:257867792
TEST=meson build && cd build && meson test

Change-Id: Id841ea6956ab4cd67836052ee8d256bf70f80a27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3952779
Reviewed-by: Timothy Loh <timloh@chromium.org>
Reviewed-by: kokoro <noreply+kokoro@google.com>
Tested-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Sophia Lin <sophialin@google.com>