shill: Update WiFi.TimeFromRekeyToFailureSeconds to use Timer
Network.Shill.WiFi.TimeFromRekeyToFailureSeconds was being triggered
when the connection becomes unreliable shortly after a rekey attempt.
This can happen organically within the same wifi service when a
device roams to a different BSSID within the same ESSID.

The metric was rewritten to have all the timing/reporting logic handled
in metrics rather than in the wifi module. This will allow the rekey
timer to be reset if the BSSID changes. This was also done as the
original unit test associated with this metric was removed as it was
only testing the functionality of the wrapper. By moving the logic to
metrics with the use of a timer, more relevant behavior was mockable and
actual edge cases can be covered in the unit tests.

Add more unit tests.

BUG=b:250079464
TEST=FEATURES=test emerge-octopus shill passes
TEST=modify shill to emit unreliable connection event when AP
disconnected, then ran tast run $DUT wifi.PSK and ensured metric was
being triggered when AP was manually killed w/ pkill hostapd after rekey
attempts.

Change-Id: Iefabccbcae530ee261a6268c3d2883fe96939b75
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3989543
Tested-by: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Kaidong Wang <kaidong@google.com>
Commit-Queue: Billy Zhao <billyzhao@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>