vtpm: provide required TPM capability properties
According to the spec these commands should be available even when the
tpm is in failure mode, which some tools and libraries use to detect if
a tpm is a tpm2 device in the first place.

This also updates the test querying for "unsupported" commands.
According to TCG "Trusted Platform Module Library Part 3: Commands"[1] in 30.2.1
if the provided property doesn't exist we should provide the next
higher-numbered one. If there are no more properties we simply return
successfully with an empty list and with the `has_more` value set to NO.

[1]: https://trustedcomputinggroup.org/wp-content/uploads/TPM-Rev-2.0-Part-3-Commands-01.38.pdf

BUG=b:254145148
TEST=unit tests; deployed to brya device and ensured properties show up
under bruschetta with `sudo tpm2_getcap properties-fixed`.

Change-Id: I728fb5cf1a85841c8328de1163252eb64df1715c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3988335
Auto-Submit: Mike Gerow <gerow@google.com>
Commit-Queue: Mike Gerow <gerow@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Mike Gerow <gerow@google.com>
Reviewed-by: Leo Lai <cylai@google.com>