secanomalyd: sets a cap for memfd baseline metric
Limits the number of times the daemon invokes the UMA endpoints for
recording a baseline metric for the memfd anomaly, which is when the
daemon detects a successful memfd_create syscall. Since memfd_create is
the baseline condition, the exact number of occurrences is not of
interest once there are enough occurrences to label the metric as
normal. By not reporting further occurrences we can avoid potential
performance hits as we stop queuing and uploading the metric.

BUG=b:257093852
TEST=All secanomalyd tests pass.
TEST=Secanomalyd stops reporting memfd_create after the cap is reached.

Change-Id: I02ebc6e14818a8b189c61fcccf4312842510d72c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3995325
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Aashay Shringarpure <aashay@google.com>
Reviewed-by: Jeff Xu <jeffxu@google.com>
Tested-by: Shervin Oloumi <enlightened@chromium.org>