power: Add delay when retrying to refresh power status.
Adds a small delay (with backoff) between each call to refresh the power
supply. This was suggested in the original cl that added retries, but
the theory at the time is that it was not required.

It is not certain that this is the correct fix for RefreshImmediately
returning false after 3 successive calls, but until we get better
logging with these crashes then we can speculate that introducing a
small delay could be a fix, and can't make it worse.

BUG=b:239908310
TEST=CQ
TEST=Build and deploy dump_power_status and power_supply info to DUT
TEST=and run command successfully.

Change-Id: I4b640322aaab353c144d0b51a457a8962b4c2b89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3990024
Tested-by: Stuart Langley <slangley@chromium.org>
Commit-Queue: Stuart Langley <slangley@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Auto-Submit: Stuart Langley <slangley@chromium.org>