power: Adjust the behaviour of the keyboard backlight toggle call
The keyboard backlight controller currently has two dimensions of
user-adjustable settings:

    * The brightness of the backlight

    * If the backlight is turned on or off

One problem with having these two dimensions if there are multiple
"backlight is off" states, which can be confusing to users. One such
example is if the user decreases the keyboard backlight brightness to
zero, and then presses the "toggle keyboard backlight" button.
Currently, this will result in the controller turning the backlight off,
despite it already being off.

This CL adjusts the behaviour of the "backlight toggle" call to simply
be a shortcut for setting the brightness to zero, or back to its
previous non-zero state. This means we no longer have the two dimensions
of settings (brightness and on/off), but just a single dimension
(brightness).

BUG=b:212618906
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/power_manager`
TEST=Manually tested on a DUT with https://crrev.com/c/3997734, ensuring
     the backlight toggles each time the keyboard button is pressed,
     even when brightness is manually adjusted to zero.

Change-Id: I6c29b3dedee29a36edba2ae979fe13333ef3c60a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4001364
Tested-by: David Greenaway <dgreenaway@google.com>
Reviewed-by: Phil Kiossoglou <pkiossoglou@chromium.org>
Reviewed-by: Mengqi Guo <mqg@chromium.org>
Commit-Queue: David Greenaway <dgreenaway@google.com>