login: Experiment for Chrome generated owner keys
Implement StartSessionEx that takes an additional argument
`chrome_owner_key`. If it's false, then StartSessionEx will behave
the same as the existing StartSession. If it's true,
session_manager will not generate the owner key and will expect
Chrome to do that and to install it together with the first
device policies. This will allow to gradually rollout the change
as an experiment.

Additionally update when and which operations with the owner keys
are allowed (the SessionManagerImpl::GetKeyInstallFlags method).

BUG=b:244407123
TEST=Build Chrome that sends the new message and manually test
  that session manager doesn't generate the key when
  chrome_owner_key==true.

Change-Id: I324817b712cf3bab58322818960bf26940ab18e7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4000400
Tested-by: Michael Ershov <miersh@google.com>
Reviewed-by: Pavol Marko <pmarko@chromium.org>
Commit-Queue: Michael Ershov <miersh@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>