shill: Convert some SLOGs to LOG(INFO)
We don't have sufficient logs in device deregister path to
debug the crash in b/230401667. And issue is not reproducible
locally by enabling verbose logging. Hence changing few SLOGs
to LOG(INFO) so that it is available in /var/log/net.log
when crashmon reports the crash in subsequent builds.

BUG=b:230401667
TEST=Verified logs are visible without verbose logging

Change-Id: I2194cc8e278cbd2adf6bff7091dbc4eb6ba8f12f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3993829
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Ujjwal Pande <ujjwalpande@google.com>
Commit-Queue: Ujjwal Pande <ujjwalpande@google.com>