faced: Fix use after free bug
Previously due to the way the CrosCameraClient is designed, there's
potential for authentication or enrollment to complete while the
callback that's passed to the CameraHAL is still active. This causes an
asan failure. This CL fixes this bug by only proceeding with
authentication and enrollment only when the CameraHAL has been notified
to stop captures.

BUG=b:255460130
TEST=Run `face_cli_test` with asan with 100 instances in parallel and
confirm that the errors are not longer present.

Change-Id: Iaa48f69b71e5e1c7fb308e731b87f4559bf893fa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3995646
Commit-Queue: David Greenaway <dgreenaway@google.com>
Auto-Submit: Victor Wong <vwon@google.com>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Tested-by: Victor Wong <vwon@google.com>