metrics: Add fstatfs to memd seccomp filter
fstatfs is used dynamically load a shared object during runtime,
to verify that the preloaded object is executable. At present,
memd never needs to do this. However, if we move to another allocator,
e.g. scudo or partition alloc, we'll need to preload those
malloc symbols from a shared library.

This CL adds fstatfs(64) accordingly to memd so that it won't
crash immediately. We need both versions of fstatfs for 32bit arm
because of trogdor.

The crash doesn't seem consistent, but we've spotted it in
/var/spool/crash and /var/log/audit. Unsure why it's not
consistent.

BUG=b:255373523, b:248085566
TEST=Ran on volteer device, ran on nami device

Change-Id: I9c7fcd7ebd816a6ed554894427e21d7f684101aa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4018393
Reviewed-by: Ian Barkley-Yeung <iby@chromium.org>
Commit-Queue: Jordan Abrahams-Whitehead <ajordanr@google.com>
Tested-by: Jordan Abrahams-Whitehead <ajordanr@google.com>
Reviewed-by: Miriam Zimmerman <mutexlox@chromium.org>