os_install_service: Remove support for old lsblk version
At the time this code was written, the lsblk version on CrOS was
2.32. In that version, JSON fields got encoded as strings even for data
that should be a bool or a number. CrOS now has lsblk 2.36.2, which
encodes bools and numbers in the appropriate type instead of a string,
so we can drop support for the old version of lsblk.

Also add an `Eq` derive to make clippy happy.

BUG=b:257078794
TEST=cros_workon_make --install --test os_install_service
TEST=Test OS install in a VM, make sure it succeeds and result boots

Change-Id: I36418f6fb800800e89fe880fe5987c91715fe25c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4000241
Reviewed-by: Steve Kendall <skend@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>