os_install_service: Avoid installing to USB HDDs
lsblk has two similar properties related to whether a device is
removable: "rm" and "hotplug". Previously we used the "rm" property,
which is not broad enough. For example, it does not mark large USB HDDs
are removable. The "hotplug" property fixes this issue.

Note that removable drives are not fully excluded as valid targets,
non-removable drives are just preferred (see
`choose_destination_device_path` for details). So this change will never
cause install to fail with "no target disk found" on a machine that
would previously have found a target disk, but it should do a better job
of prioritizing internal drives.

BUG=b:254501715
TEST=cros_workon_make --install --test os_install_service
TEST=Run OS install on hardware with a big USB HDD attached.
TEST=Make sure that the smaller internal drive is chosen over the USB drive.

Change-Id: I5f412210f0918ead083ba31e1fd3c0396576dfd2
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4000242
Tested-by: Nicholas Bishop <nicholasbishop@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Reviewed-by: Steve Kendall <skend@google.com>