vm_tools: cros_im: Add Qt5 IME plugins
In its current status, Qt5 plugin works under both XWayland and
native wayland, Qt6 plugin also works but are not being built as
we are building this package for debian bullseye and Qt6 is not
available in this version of debian.

Surrounding text related functionality is not implemented, so
languages requiring this feature can't be used.

BUG=b:232048947
TEST=featherpad for Qt5, can type zh_CN

Change-Id: Ia918d9b8faef8035212116565e6053592d05baac
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4000072
Reviewed-by: kokoro <noreply+kokoro@google.com>
Commit-Queue: Zihan Chen <zihanchen@google.com>
Reviewed-by: Timothy Loh <timloh@chromium.org>
Tested-by: kokoro <noreply+kokoro@google.com>