cryptohome: Switch PrepareAuthFactor over to returning tokens
Instead of returning a status, and having the active factors being
tracked by username (in the fingerprint service) and by type (in auth
session, which must terminate them manually) this instead replaces the
tracking with a token object which terminates the factor upon
destruction.

This also lets us cut down the auth block utility API as it no longer
needs to expose Terminate functionality, as individual factors can do
that themselves in their token implementations.

As a side effect, this fixes a bug where where an asynchronous failure
in preparing an auth factor would still leave it to be considered
active, potentially resulting in a factor being terminated by the wrong
session, if you sequenced things in such a way as to end up with two
sessions that both thought a factor was active. This sort of error isn't
really possible with the token approach because the uniqueness of the
token prevents accidental "double ownership" implicitly.

BUG=b:257124612
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ia2ab2d7d726ff5e59a3c2e1d8f01d17f617cbeda
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4004665
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>