chromeos-config: crosid: Support printing a single value
The typical output of crosid is so that it can be sourced by a shell
script and expose variables.

For a script using just a single variable (which is the typical case),
it's simpler and doesn't upset shellcheck as much to have crosid be
able to provide values which aren't shell-quoted, and only the single
value.

For example, this enables doing:

  local sku_id="$(crosid -f SKU)"

... instead of:

  eval "$(crosid)"
  local sku_id="${SKU}"

BUG=b:233939831
TEST=provided unit tests pass

Change-Id: Ica8baa085f30bc1efb115e1388daef36f77343bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4004667
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>