diagnostics: Implement HasPrivacyScreen method
This CL implements the HasPrivacyScreen method and puts a flag in each
board project so that privacy screen routine can be listed on the board
that has a built-in privacy screen.

Such method is not implemented by a software detecting a built-in
privacy screen because we need the routine be listed even if the
software or the driver is broken.

BUG=b:239892143
TEST=Build diagnostics project

Cq-Depend: chromium:3995697
Change-Id: Idb85eeb19922387625f26621c46c8c959a2e80a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3915663
Reviewed-by: Kerker Yang <kerker@chromium.org>
Tested-by: Wen-Ting Tseng <wdzeng@chromium.org>
Commit-Queue: Wen-Ting Tseng <wdzeng@chromium.org>