power: Make SimpleAlarmTimer take a clock in its constructor.
SimpleAlarmTimer currently takes no arguments to its static Create
method and creates timers based on the CLOCK_REALTIME_ALARM clock
type.

It should also be able to create timers based on the
CLOCK_BOOTTIME_ALARM clock type.

This change updates the SimpleAlarmTimer::Create method to take
a single argument with the clock_id of the type of alarm to be created.
It defaults to CLOCK_REALTIME_ALARM so as not to disrupt other invokers
of this class.

Valid clock types are now either of the alarm clock types - that is
CLOCK_REALTIME_ALARM and CLOCK_BOOTTIME_ALARM.

BUG=b:69759087
TEST=emerged, ran unit tests

Change-Id: If496546c985436d8709b69d27d937541cde86bf7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4008938
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Phil Kiossoglou <pkiossoglou@chromium.org>
Reviewed-by: David Greenaway <dgreenaway@google.com>
Commit-Queue: Phil Kiossoglou <pkiossoglou@chromium.org>