secagentd: Add an Image checksum cache
The image checksum cache amortizes the cost of hashing process
executable images across process invocations.
The cache uses inodes to look up cached information. Inode mtime and
ctime are mixed in to verify freshness just in case the executable is
modified. "This should never happen on CrOS". Similar to the process
cache, this too is implemented as a LRU cache of a fixed size.

Sha256 checksums are calculated using OpenSSL.

BUG=b:242604609
TEST=All unit tests pass
TEST=Manually verified some process exec protos vs. binary checksums on
TEST=a amd64-generic VM.

Change-Id: Ie6132f12e2ba10d0fc4429a1eb62c31effbf3773
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4015051
Reviewed-by: Ryan Borzello <rborzello@google.com>
Tested-by: Aashay Shringarpure <aashay@google.com>
Commit-Queue: Aashay Shringarpure <aashay@google.com>
Reviewed-by: Jeff Xu <jeffxu@google.com>
Reviewed-by: Ben Scarlato <akhna@google.com>