camera: Fix the crash and deadlock issues when minimizing CCA after taking a portrait photo
There is a deadlock case between reprocess_effect_thread_ and
camera_device_ops_thread_. Remove future.Get() waiting in a thread to
ensure the non-blocking thread model.

We call the Close() function when minimizing CCA, and a crash will
happen when we want to complete the pending reprocess request. Add a
check to ensure we execute process_reprocess_request only if the camera
is open.

BUG=b:247488022
TEST=tast run kukui camera.HAL3PortraitMode
TEST=no crash occurs when minimizing CCA in portrait reprocessing

Change-Id: I5c1829a2bae67a2a17ff62d43c80d33312dcabd6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4008042
Tested-by: Juliana Chang <julianachang@chromium.org>
Reviewed-by: Ren-Pei Zeng <kamesan@chromium.org>
Commit-Queue: Juliana Chang <julianachang@chromium.org>