vtpm: support algorithms capability query.
This is mainly to provide better support for testing with chromeos'
library.

Specifically, `trunks::TpmStateImpl` treats such a capability as
critical information and aborts if the query fails. if stops developers
from using the existent codes to validate vTPM.

Also, potentially the consumer of vTPM in production might be also
interested in what algorithms are supported.

Since technically all the algorithm specifier that are forwarded to the
host TPM can be supported from the consumer POV, currently we forward
the answer by the host TPM.

BUG=b:258203810
TEST=`trunks_client --vtpm --tpm_version`

Change-Id: I89f1e9ae2bd97546af0207b3361e5c4e9b4872b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4018672
Commit-Queue: Leo Lai <cylai@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Leo Lai <cylai@google.com>