camera: fake_hal: Get optional dict/list value without warning.
Previously, the GetValue has an optional third argument for default
value, which is used to get optional primitive types without warning
when the value is not present, and this pattern doesn't work well for
dict and list value. Change it to two separate functions GetValue and
GetRequiredValue, which both return std::optional but the GetValue by
default doesn't warn when the value is missing.

BUG=b:236782829
TEST=compile

Change-Id: Ie4512f516a91bd31b689f41df5c7040619ed9176
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4016139
Tested-by: Pi-Hsun Shih <pihsun@chromium.org>
Commit-Queue: Pi-Hsun Shih <pihsun@chromium.org>
Reviewed-by: Shik Chen <shik@chromium.org>