shill: refactor session tag state handling
Refactor the logic that tests if the session tag is in a "good" state
before emitting the WiFi structured metrics. Instead of more or less
copy/pasting almost the same bit of code everywhere, create a private
method. That will be useful as we add more events that all need to
report the session tag.

BUG=None
TEST=cros-workon-${BOARD} start shill
TEST=FEATURES=test emerge-${BOARD} shill

Change-Id: Ieb07b039cf34a6d523719f01594dbf4a2803775a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4019315
Reviewed-by: David Ruth <druth@chromium.org>
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>