chromeos-config: Migrate cros_config_mock to use crosid for identity scanning
The only place this script is used is for the factory process, in the
following fashion:

  cros_config_mock --sku-id SKU [--custom-label-tag TAG] /path property

(See src/platform/factory/py/test/pytests/update_cbi.py)

Based on the discussion in CL:4006562, we need to keep this script
around longer term to support factory usage, and the resolution of
b/152291015 doesn't allow us to delete it yet.  Thus, we should (1)
have less logic in an untested shell script and (2) unify that logic
with crosid internally.  By adding two new flags to crosid to support
the overrides desired, we can support doing that.

Finally, we can remove flags to this script the factory is not
currently using.

BUG=b:152291015
TEST=crosid unit tests pass
TEST="cros_config_mock / name" on brya

Change-Id: Ib69c42ab06ca9e37cf845913cb67ebfd6ee07a77
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4021725
Tested-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Raul Rangel <rrangel@chromium.org>
Commit-Queue: Jack Rosenthal <jrosenth@chromium.org>