hiberman: No-op edit to UswsuspUserKey to match kernel upstream
The kernel folks wanted me to get rid of the pad member of struct
uswsusp_user_key. Due to the required 8-byte alignment of the
meta_size member, this doesn't actually change the size or layout of
the struct, it just makes that pad member an implicit compiler field,
rather than a named field.

Make the same change here in Rust. I was originally worried I needed
this and the repr(packed) because I'd otherwise run afoul of
soundness rules related to undefined values (which data between struct
members is) and transmuting between bytes and structs. But we're careful
about how we construct the struct from a u8 array (using ::default() and
then filling in individual fields), so we don't run afoul of those rules.

BUG=b:237086416
TEST=hiberman hibernate --test-keys --dry-run, no ioctl error

Change-Id: I240ae5a7f535c12ae68af6d819e3827339c605fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4024189
Reviewed-by: Allen Webb <allenwebb@google.com>
Tested-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Evan Green <evgreen@chromium.org>