cryptohome: Increase USS version to 2
UserSecretStash is enabled on ToT for M109 by default,
then disabled on the server side by marking the USS v=1
as invalid (i.e. last_invalid=1) as a last minute decision
to disable the feature for M109.
To enable UserSecretStash for M110 and not enable for M109
we need to increase USS version in the code.

BUG=b:257290475
TEST=Unit tests pass.

Change-Id: I82a989a2d68e7e2be06d31b8ce64217277c59710
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4021109
Reviewed-by: Howard Yang <hcyang@google.com>
Commit-Queue: Betul Soysal <betuls@google.com>
Tested-by: Betul Soysal <betuls@google.com>