cryptohome: Get rid of some test friend declarations
Came across these when I was doing my big refactoring of all the
verifier code, meant to send out a cleanup. Some of the declarations are
just unnecessary now, and the others are for some functions that can
just be in the public interface.

BUG=none
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Icf7bfb78d68c0586028a673e44d7fbb41ae62396
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4020943
Reviewed-by: Thomas Cedeno <thomascedeno@google.com>
Auto-Submit: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>