cryptohome: Extract the AuthFactor map in AuthSession to a class
This change makes a standalone AuthFactorMap class, similar to the
UserSessionMap we created a while ago. Currently we just expose the raw
std::map from AuthSession, which makes it somewhat painful to change.
Extracting this into a standalone class with a narrower interface should
make it a little easier to make modifications.

The intent is then to enhance this to allow the underlying storage that
each factors was loaded from (i.e. vault keyset or USS) to be tracked as
well.

BUG=b:258531076
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I6585b576e85856693be059ff212c9a5fd4a45f88
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4024186
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>