missive: Set, propagate and handle reserved space
This is a request for opportunistic queues - matching Chromium side
crrev.com/c/4018222. Allows to specify requirement in the
ReportQueueFactory (or directly in ReportQueueConfiguration) and then
pass it to the Storage, so that if there is less space than it should be
reserved, Write will be rejected.

This is a replica of what was done on Chromium size to support opportunistic queues. With this new change the ReportQueueConfiguration
may be set on the Chromium, and will be handled on the daemon - since the respective setting is attached to `Record` that is then passed
to the daemon enqueue entry point.

ChromeOS-side clientlib also has respective capability now too.

BUG=b:256918709
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: Ib97d54e74a2107077d404913570001bce845800b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4023499
Reviewed-by: Hong Xu <xuhong@google.com>
Auto-Submit: Leonid Baraz <lbaraz@chromium.org>
Tested-by: Leonid Baraz <lbaraz@chromium.org>
Commit-Queue: Hong Xu <xuhong@google.com>