cryptohome: Add storage type to AuthFactorMap
This changes the Add to require you to also specify the underlying
storage type. It doesn't currently export this information in any of the
lookups because we don't yet have anything that would actually make use
of it.

In the future the lookups that require the info will need to extend the
API to export it, but coming up with the most effective way of doing
that would be easier with a specific use case.

BUG=b:258531076
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I92024c14039df7dc8eda1b6c9ec435aebc5d5ef9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4020833
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>