power: Only report AmbientLightOnResume UMA when lid is open
Currently the AmbientLightOnResume UMA is reported any time the device resumes from suspend. This could occur when the lid is closed, such as when using an external monitor, or during dark resume.

This change modifies the internal backlight controllers reporting of the lux reading from the ambient light sensor to only occur if the lid is open. The intent of the UMA is better met by ignoring the readings when the lid is closed.

BUG=b:256693273
TEST=`cros_workon_make --board=$BOARD --test chromeos-base/power_manager`

Change-Id: Id74990a1916b33d5caa2529be69f78f4b6da8eb0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4024885
Tested-by: Ben Kersten <bkersten@google.com>
Commit-Queue: Ben Kersten <bkersten@google.com>
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
Reviewed-by: Stuart Langley <slangley@chromium.org>