oobe_config: Add FileHandler for filesystem access
Add a class to encapsulate file system usage that has the root path as
member variable. Usually, root path simply is "/" but it can be adapted
for tests. This way, tests can pretend to be using the filesystem while
writing under some temp directory.

The rollback_helper class still accesses filesystem directly. I am
planning to move this partially to tmpfile configs and partially to the
FileHandler in a follow-up.

BUG=b:234826714
TEST=1)FEATURES=test emerge-brya oobe_config
     2)USE="asan fuzzer" emerge-brya oobe_config
     3)tast run <dut> policy.EnterpriseRollbackInPlace
     4)touch /mnt/stateful_partition/.save_rollback_data
       echo "fast safe keepimg" >
       /mnt/stateful_partition/factory_install_reset
       reboot

Change-Id: I56f124c12d863d7dffa5cb24f2d785b2badcfd68
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/3991049
Reviewed-by: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Roland Bock <rbock@google.com>
Tested-by: Miriam Polzer <mpolzer@google.com>
Commit-Queue: Miriam Polzer <mpolzer@google.com>