cryptohome: Make use of auth factor storage type for decisions
For individual operations (e.g. AuthenticateAuthFactor) where we have
separate VK and USS paths this changes the decision as to which path to
use, to use the storage type associated with the specific factor instead
of more global information.

This also completely eliminates the user_has_* variables and replaces
them with checks of the auth factor map. In some cases we can just do a
simple empty/not-empty check, and in others we use new HasStorageType
functions on the map that checks if the map contains any keys with a
particular storage type.

BUG=b:258531076
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I08bc9a072fb78144de3b0b1067bcc74b324acffb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4031367
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Betul Soysal <betuls@google.com>
Commit-Queue: Betul Soysal <betuls@google.com>
Reviewed-by: Maksim Ivanov <emaxx@chromium.org>