biod: Add a function to get hardware Errors
Biod can detect whenever a fingerprint sensor is not responding due to a
bad hardware. Add a private method `GetHwError` in CrosFpDevice to
return any hardware errors in the fingerprint sensor. This method can be
called at the start of the Auth or Enroll Session to provide feedback to
UI.

BUG=b:230478501
TEST=cros_workon_make --board=hatch --test biod

Change-Id: I0c1a8d67acd6f2693fb2ec8c9812338fb94a67b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4029330
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Firas Sammoura <fsammoura@google.com>
Tested-by: Firas Sammoura <fsammoura@google.com>