shill: rename Nl80211StaInfo to StationStats
Follow-up of https://crrev.com/c/4024372, change the name of the struct
to reflect the fact that it's not a direct mirror of the
nl80211_sta_info struct. It will contain some of the information found
in nl80211_sta_info, but not only. It's meant to be a struct that
contains all the useful statistics of the WiFi link

BUG=b:230525092
BUG=b:239864299
TEST=cros-workon-${BOARD} start shill
TEST=FEATURES=test emerge-${BOARD} shill

Disallow-Recycled-Builds: scarlet-cq
Change-Id: I0259343df7feb45a2c6006e9230dcec1f2770bc6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4033713
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>
Reviewed-by: David Ruth <druth@chromium.org>