shill: use StationStats rather than KeyValue store
When shill receives nl80211 STA information from the driver and parses
it, it now uses the specialized struct WiFiLinkStatistics::StationStats
rather than storing everything in a key value store that needs to be
queried and converted back to a more usable format.
The StationStats struct can still be converted to a key value store when
required to expose some of the properties over D-Bus.

BUG=b:230525092
BUG=b:239864299
TEST=cros-workon-${BOARD} start shill
TEST=FEATURES=test emerge-${BOARD} shill

Disallow-Recycled-Builds: scarlet-cq
Change-Id: Ifdb579075e02d23ab2d37fc79fea603615992c97
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4033717
Reviewed-by: Matthew Wang <matthewmwang@chromium.org>
Commit-Queue: Nicolas Norvez <norvez@chromium.org>
Tested-by: Nicolas Norvez <norvez@chromium.org>